backend/requirements.txt
Django
psycopg2
djangorestframework
djangorestframework-jwt
django-cors-headers